refactor(rpc): reuse native multisig config loading
